feat: full audit remediation — 4 waves close 25+ findings
Waves A-D applied a thorough static + security audit across the workspace
and remediated every CRITICAL / HIGH / MEDIUM finding.

NEW HOOKS — block escape vectors the previous build missed
  proc_guard:   NtCreateProcess + NtCreateProcessEx (legacy syscalls bypassing
                NtCreateUserProcess hook)
  com_guard:    RoGetActivationFactory + RoActivateInstance (WinRT bypass via
                Windows.System.Launcher.LaunchUriAsync etc.) with HSTRING
                decode + class-name denylist
  system_guard: NtCreateSymbolicLinkObject (object-manager symlink forge),
                NtLoadDriver / NtUnloadDriver (BYOVD: dbutil_2_3, RTCore64,
                etc.) returning STATUS_PRIVILEGE_NOT_HELD
  shell_guard:  ShellExecuteW + ShellExecuteExW (ms-windows-store:,
                ms-settings:, shell:AppsFolder\<AUMID>, runas:) — new module
  fs_metadata_guard: NtSetEaFile (NTFS EA covert persistence channel)

LAUNCHER HARDENING
  sandbox.rs:   refuse 32-bit / WoW64 children via IsWow64Process2 — our
                x64 ntdll hooks never see 32-bit syscall paths
  pipe_server:  extend PERSISTENCE_DENY_SUFFIXES from 6 → 25 (Run/RunOnce,
                Winlogon, Drivers32, COM CLSID InprocServer32, file
                associations, LSA Notification Packages, Office Trusted
                Locations, ...)
  pipe_server:  semaphore-cap concurrent handler tasks at 128 to bound
                blocking-pool consumption under hostile load
  pipe_server:  delete dead policy_net_is_denied stub; net rules are
                enforced by WFP, not the policy table

FS HOOK CORRECTNESS
  fs_hooks:     reparse-point CREATE-side block (FILE_OPEN_REPARSE_POINT +
                CREATE disposition); existing FSCTL post-create block was
                only half the story
  fs_hooks:     reject any non-null EaBuffer on NtCreateFile (NTFS Extended
                Attributes were unhooked, a covert-storage primitive)
  fs_hooks:     Mock-overlay materialization is idempotent — was being
                rewritten on every open, racing concurrent readers
  fs_hooks:     Cow with missing overlay path → STATUS_ACCESS_DENIED instead
                of falling through to the real filesystem (fail-closed)
  fs_hooks:     extract HookedAttrs helper for the
                (Vec<u16>, UNICODE_STRING, OBJECT_ATTRIBUTES) lifetime
                invariant — was duplicated 8× with implicit ordering
  fs_hooks:     TOCTOU double-fetch defense — copy ObjectName.Buffer into
                hook-owned memory before passthrough kernel call
                (attacker thread cannot mutate the path between our check
                and the kernel's read)

PATH NORMALIZATION
  hooks.rs:     strip trailing dots/spaces per segment before classification
                — NTFS canonicalizes them away, our denylist did not, so
                'C:\sandbox\.winrsbox.' bypassed the .winrsbox hide
  hooks.rs:     ASCII-only lowercase everywhere security-sensitive; pin tests
                lock the Unicode U+0130 / German ß folding mismatch

POLICY
  policy/dev:   reclassify \Device\KsecDD, \Device\CNG, \Device\CNGBootInfo
                as Unknown (deny by default) — LSA / DPAPI master-key
                extraction surface had been SystemQuery (read-OK)
  alpc_guard:   add appinfo, appxsvc, appx, dcomlaunch, pchsvc,
                terminalserver, iiscertobj to ALPC denylist

CODE QUALITY / REFACTORING
  hook crate:   consolidate STATUS_ACCESS_DENIED + STATUS_PRIVILEGE_NOT_HELD
                from 9 duplicate const declarations to a single canonical
                one in hooks.rs; pin tests on the hex values
  hook crate:   split hooks.rs (1523 LOC) into ipc_client.rs + fs_hooks.rs
                + slimmer hooks.rs (decide, orchestration, install_hooks)
  hook crate:   delete orphan hook/src/reg_overlay.rs (-204 LOC) — was
                never read anywhere, RegWrite IPC now routes directly
                through the launcher's authoritative on-disk overlay
  launcher:     split main.rs (1653 LOC) into inject.rs + pipe_server.rs +
                sandbox.rs + slimmer main.rs (CLI + dispatch only)
  launcher:     drop unused windows-targets dep (pulled transitively by
                windows-0.61)
  policy:       split lib.rs into policy_impl.rs + decide.rs + registry.rs;
                lib.rs becomes a thin facade with public re-exports
  policy:       narrow pub mod decide / pub mod registry to pub(crate) —
                external callers go through the re-exports
  policy:       dedup ensure_lower (was defined twice with identical bodies)

RUST-INTEL FIXES (post-audit)
  com_guard:    extern "C" → extern "system" on COM type aliases (would
                corrupt stack on x86 builds; no-op on x64)
  ui_guard:     transmute(addr) → transmute(addr as usize) (raw-ptr to
                fn-ptr direct transmute is technically UB)
  mitigations:  DEP_ENABLE / DEP_ATL_THUNK_ENABLE / SEHOP_ENABLE were at
                bit positions 36/40/44 — same bits as PROHIBIT_DYNAMIC_CODE
                etc. Fixed to canonical 0/1/2 per Microsoft docs
  ipc_client:   fail threshold 10 → 3; passthrough → deny while under
                threshold (audit T1/T2: was effectively fail-open for first
                9 failures); pin test + THREATMODEL.md doc sync
  reg_overlay:  is_key_deleted exact-match-or-subkey (was over-matching
                lookalike siblings: delete_key('HKLM\test') reported
                'HKLM\testlab\v' as deleted)
  reg_overlay:  tombstones Vec<String> → HashSet<String> (O(n) → O(1)
                lookup + dedup on repeated delete_value)
  inject.rs:    SAFETY block on extract_child_exe documenting the 0x60
                offset assumption + length sanity bound
  inject.rs:    pin tests for the intentional remote_buf leak (regression
                guard: a future VirtualFreeEx after NtQueueApcThread would
                crash child silently)
  + 83 SAFETY comments added to previously-bare unsafe blocks across the
  guard modules

TESTS
  +63 new hook unit tests   (73 → 136 passed)
  +4  new policy unit tests (227 → 231 passed)
  +26 new winrsbox bin tests (persistence-denylist coverage etc.)
  +1  e2e: two_concurrent_children_both_succeed — one launcher,
      two concurrent sandboxed writers, both overlays materialized
      independently with no IPC interference
  +pin tests for IPC_FAIL_THRESHOLD, STATUS_* constants, intentional leaks,
   process_tracker PID-reuse contract, PS_ATTRIBUTE_LIST edge cases

BENCHES
  hook/benches/path_traversal.rs — adversarial inputs for check_path_traversal
    (32k path, deep nesting, GLOBALROOT, late 8.3 tilde, .winrsbox match).
    Worst-case ~40us vs ~1.6us normal — pinned to catch future regressions
  hook/benches/decide_full.rs — cache-hit path (~101 ns), IPC half left to
    future integration-bench harness (criterion + self-terminate IPC don't mix)

+++ winrsbox/hook/src/ui_guard.rs
@@ -60,6 +60,9 @@ static HOOK_SEND_MESSAGE_A:    OnceLock<GenericDetour<FnSendMessageA>>    = Once
 /// Returns true when `hwnd` is a window owned by a process **other** than us.
 /// For cross-process HWNDs we deny PostMessage/SendMessage. Own-process
 /// windows still work normally.
+///
+/// # SAFETY
+/// `hwnd` may be any HWND value; null is handled internally.
 unsafe fn is_foreign_hwnd(hwnd: HWND) -> bool {
     if hwnd.is_null() { return false; }
     let mut pid: DWORD = 0;
@@ -80,28 +83,33 @@ fn report_and_kill(api: &str) -> ! {
         ipc_log(ipc::LogLevel::Warn,
             format!("INPUT-INJECT DENY: {api} — terminating process"));
     }
-    // Mirror memory_guard kill pattern: terminate self, then loop to prevent return.
+    // SAFETY: TerminateProcess on own handle is always valid; intentional self-termination.
     unsafe {
         winapi::um::processthreadsapi::TerminateProcess(
             winapi::um::processthreadsapi::GetCurrentProcess(),
             0xC000_0005,
         );
     }
+    // SAFETY: Sleep(INFINITE-like) after TerminateProcess — defensive loop in case terminate is async.
     loop {
         unsafe { winapi::um::synchapi::Sleep(1000) };
     }
 }
 
+// SAFETY: Called by detour2 dispatcher with user32!SendInput ABI.
 unsafe extern "system" fn hook_send_input(n: UINT, inputs: *mut INPUT, sz: i32) -> UINT {
     let Some(_g) = anti_rec::enter() else {
+        // SAFETY: detour2 trampoline matches FnSendInput ABI.
         return HOOK_SEND_INPUT.get().unwrap().call(n, inputs, sz);
     };
     let _ = _g; // guard held until report_and_kill diverges
     report_and_kill("SendInput")
 }
 
+// SAFETY: Called by detour2 dispatcher with user32!keybd_event ABI.
 unsafe extern "system" fn hook_keybd_event(b: u8, s: u8, f: DWORD, ex: usize) {
     let Some(_g) = anti_rec::enter() else {
+        // SAFETY: detour2 trampoline matches FnKeybdEvent ABI.
         HOOK_KEYBD_EVENT.get().unwrap().call(b, s, f, ex);
         return;
     };
@@ -109,8 +117,10 @@ unsafe extern "system" fn hook_keybd_event(b: u8, s: u8, f: DWORD, ex: usize) {
     report_and_kill("keybd_event")
 }
 
+// SAFETY: Called by detour2 dispatcher with user32!mouse_event ABI.
 unsafe extern "system" fn hook_mouse_event(f: DWORD, x: DWORD, y: DWORD, d: DWORD, ex: usize) {
     let Some(_g) = anti_rec::enter() else {
+        // SAFETY: detour2 trampoline matches FnMouseEvent ABI.
         HOOK_MOUSE_EVENT.get().unwrap().call(f, x, y, d, ex);
         return;
     };
@@ -118,16 +128,20 @@ unsafe extern "system" fn hook_mouse_event(f: DWORD, x: DWORD, y: DWORD, d: DWOR
     report_and_kill("mouse_event")
 }
 
+// SAFETY: Called by detour2 dispatcher with user32!BlockInput ABI.
 unsafe extern "system" fn hook_block_input(fblock: BOOL) -> BOOL {
     let Some(_g) = anti_rec::enter() else {
+        // SAFETY: detour2 trampoline matches FnBlockInput ABI.
         return HOOK_BLOCK_INPUT.get().unwrap().call(fblock);
     };
     let _ = _g;
     report_and_kill("BlockInput")
 }
 
+// SAFETY: Called by detour2 dispatcher with user32!SetCursorPos ABI.
 unsafe extern "system" fn hook_set_cursor_pos(x: i32, y: i32) -> BOOL {
     let Some(_g) = anti_rec::enter() else {
+        // SAFETY: detour2 trampoline matches FnSetCursorPos ABI.
         return HOOK_SET_CURSOR_POS.get().unwrap().call(x, y);
     };
     let _ = _g;
@@ -141,10 +155,13 @@ unsafe extern "system" fn hook_set_cursor_pos(x: i32, y: i32) -> BOOL {
 // bypassing our user-mode patch. SendInput-class hooks (the actual Win+R
 // vector) work reliably. Job UI restrictions cover the rest in principle.
 
+// SAFETY: Called by detour2 dispatcher with user32!FindWindowW ABI.
 unsafe extern "system" fn hook_find_window_w(class: LPCWSTR, name: LPCWSTR) -> HWND {
     let Some(_g) = anti_rec::enter() else {
+        // SAFETY: detour2 trampoline matches FnFindWindowW ABI.
         return HOOK_FIND_WINDOW_W.get().unwrap().call(class, name);
     };
+    // SAFETY: detour2 trampoline matches FnFindWindowW ABI; same args passed through.
     let hwnd = HOOK_FIND_WINDOW_W.get().unwrap().call(class, name);
     if is_foreign_hwnd(hwnd) {
         log_soft_deny("FindWindowW", "foreign HWND");
@@ -153,10 +170,13 @@ unsafe extern "system" fn hook_find_window_w(class: LPCWSTR, name: LPCWSTR) -> H
     hwnd
 }
 
+// SAFETY: Called by detour2 dispatcher with user32!FindWindowA ABI.
 unsafe extern "system" fn hook_find_window_a(class: LPCSTR, name: LPCSTR) -> HWND {
     let Some(_g) = anti_rec::enter() else {
+        // SAFETY: detour2 trampoline matches FnFindWindowA ABI.
         return HOOK_FIND_WINDOW_A.get().unwrap().call(class, name);
     };
+    // SAFETY: detour2 trampoline matches FnFindWindowA ABI; same args passed through.
     let hwnd = HOOK_FIND_WINDOW_A.get().unwrap().call(class, name);
     if is_foreign_hwnd(hwnd) {
         log_soft_deny("FindWindowA", "foreign HWND");
@@ -165,12 +185,15 @@ unsafe extern "system" fn hook_find_window_a(class: LPCSTR, name: LPCSTR) -> HWN
     hwnd
 }
 
+// SAFETY: Called by detour2 dispatcher with user32!FindWindowExW ABI.
 unsafe extern "system" fn hook_find_window_ex_w(
     parent: HWND, child: HWND, class: LPCWSTR, name: LPCWSTR,
 ) -> HWND {
     let Some(_g) = anti_rec::enter() else {
+        // SAFETY: detour2 trampoline matches FnFindWindowExW ABI.
         return HOOK_FIND_WINDOW_EX_W.get().unwrap().call(parent, child, class, name);
     };
+    // SAFETY: detour2 trampoline matches FnFindWindowExW ABI; same args passed through.
     let hwnd = HOOK_FIND_WINDOW_EX_W.get().unwrap().call(parent, child, class, name);
     if is_foreign_hwnd(hwnd) {
         log_soft_deny("FindWindowExW", "foreign HWND");
@@ -179,12 +202,15 @@ unsafe extern "system" fn hook_find_window_ex_w(
     hwnd
 }
 
+// SAFETY: Called by detour2 dispatcher with user32!FindWindowExA ABI.
 unsafe extern "system" fn hook_find_window_ex_a(
     parent: HWND, child: HWND, class: LPCSTR, name: LPCSTR,
 ) -> HWND {
     let Some(_g) = anti_rec::enter() else {
+        // SAFETY: detour2 trampoline matches FnFindWindowExA ABI.
         return HOOK_FIND_WINDOW_EX_A.get().unwrap().call(parent, child, class, name);
     };
+    // SAFETY: detour2 trampoline matches FnFindWindowExA ABI; same args passed through.
     let hwnd = HOOK_FIND_WINDOW_EX_A.get().unwrap().call(parent, child, class, name);
     if is_foreign_hwnd(hwnd) {
         log_soft_deny("FindWindowExA", "foreign HWND");
@@ -193,26 +219,32 @@ unsafe extern "system" fn hook_find_window_ex_a(
     hwnd
 }
 
+// SAFETY: Called by detour2 dispatcher with user32!OpenClipboard ABI.
 unsafe extern "system" fn hook_open_clipboard(hwnd: HWND) -> BOOL {
     let Some(_g) = anti_rec::enter() else {
+        // SAFETY: detour2 trampoline matches FnOpenClipboard ABI.
         return HOOK_OPEN_CLIPBOARD.get().unwrap().call(hwnd);
     };
     log_soft_deny("OpenClipboard", "denied");
     0
 }
 
+// SAFETY: Called by detour2 dispatcher with user32!GetClipboardData ABI.
 unsafe extern "system" fn hook_get_clipboard_data(format: UINT) -> HANDLE {
     let Some(_g) = anti_rec::enter() else {
+        // SAFETY: detour2 trampoline matches FnGetClipboardData ABI.
         return HOOK_GET_CLIPBOARD.get().unwrap().call(format);
     };
     log_soft_deny("GetClipboardData", "denied");
     std::ptr::null_mut()
 }
 
+// SAFETY: Called by detour2 dispatcher with user32!PostMessageW ABI.
 unsafe extern "system" fn hook_post_message_w(
     hwnd: HWND, msg: UINT, wparam: WPARAM, lparam: LPARAM,
 ) -> BOOL {
     let Some(_g) = anti_rec::enter() else {
+        // SAFETY: detour2 trampoline matches FnPostMessageW ABI.
         return HOOK_POST_MESSAGE_W.get().unwrap().call(hwnd, msg, wparam, lparam);
     };
     if is_foreign_hwnd(hwnd) {
@@ -222,10 +254,12 @@ unsafe extern "system" fn hook_post_message_w(
     HOOK_POST_MESSAGE_W.get().unwrap().call(hwnd, msg, wparam, lparam)
 }
 
+// SAFETY: Called by detour2 dispatcher with user32!PostMessageA ABI.
 unsafe extern "system" fn hook_post_message_a(
     hwnd: HWND, msg: UINT, wparam: WPARAM, lparam: LPARAM,
 ) -> BOOL {
     let Some(_g) = anti_rec::enter() else {
+        // SAFETY: detour2 trampoline matches FnPostMessageA ABI.
         return HOOK_POST_MESSAGE_A.get().unwrap().call(hwnd, msg, wparam, lparam);
     };
     if is_foreign_hwnd(hwnd) {
@@ -235,10 +269,12 @@ unsafe extern "system" fn hook_post_message_a(
     HOOK_POST_MESSAGE_A.get().unwrap().call(hwnd, msg, wparam, lparam)
 }
 
+// SAFETY: Called by detour2 dispatcher with user32!SendMessageW ABI.
 unsafe extern "system" fn hook_send_message_w(
     hwnd: HWND, msg: UINT, wparam: WPARAM, lparam: LPARAM,
 ) -> isize {
     let Some(_g) = anti_rec::enter() else {
+        // SAFETY: detour2 trampoline matches FnSendMessageW ABI.
         return HOOK_SEND_MESSAGE_W.get().unwrap().call(hwnd, msg, wparam, lparam);
     };
     if is_foreign_hwnd(hwnd) {
@@ -248,10 +284,12 @@ unsafe extern "system" fn hook_send_message_w(
     HOOK_SEND_MESSAGE_W.get().unwrap().call(hwnd, msg, wparam, lparam)
 }
 
+// SAFETY: Called by detour2 dispatcher with user32!SendMessageA ABI.
 unsafe extern "system" fn hook_send_message_a(
     hwnd: HWND, msg: UINT, wparam: WPARAM, lparam: LPARAM,
 ) -> isize {
     let Some(_g) = anti_rec::enter() else {
+        // SAFETY: detour2 trampoline matches FnSendMessageA ABI.
         return HOOK_SEND_MESSAGE_A.get().unwrap().call(hwnd, msg, wparam, lparam);
     };
     if is_foreign_hwnd(hwnd) {
@@ -261,9 +299,12 @@ unsafe extern "system" fn hook_send_message_a(
     HOOK_SEND_MESSAGE_A.get().unwrap().call(hwnd, msg, wparam, lparam)
 }
 
+/// # SAFETY
+/// Must be called from install_hooks() in DllMain context with anti_rec entered.
 pub unsafe fn install() -> Result<(), Box<dyn std::error::Error>> {
     // Resolve user32 module — load if not already loaded.
     let user32_w: Vec<u16> = "user32.dll\0".encode_utf16().collect();
+    // SAFETY: FFI call to LoadLibraryW with null-terminated wide string; no outstanding borrows.
     let user32 = winapi::um::libloaderapi::LoadLibraryW(user32_w.as_ptr());
     if user32.is_null() {
         return Err("LoadLibraryW(user32.dll) failed".into());
@@ -274,7 +315,8 @@ pub unsafe fn install() -> Result<(), Box<dyn std::error::Error>> {
             let addr = winapi::um::libloaderapi::GetProcAddress(
                 user32, concat!($sym, "\0").as_ptr() as *const _);
             if !addr.is_null() {
-                let target: $ty = std::mem::transmute(addr);
+                // SAFETY: transmute of GetProcAddress result; ABI matches the hook function type $ty.
+                let target: $ty = std::mem::transmute(addr as usize);
                 let hook_ptr: $ty = $hook;
                 if let Ok(detour) = GenericDetour::<$ty>::new(target, hook_ptr) {
                     $lock.set(detour).ok();
@@ -311,6 +353,8 @@ pub unsafe fn install() -> Result<(), Box<dyn std::error::Error>> {
     Ok(())
 }
 
+/// # SAFETY
+/// Must be called from DLL_PROCESS_DETACH only.
 pub unsafe fn uninstall() {
     if let Some(h) = HOOK_SEND_INPUT.get()     { let _ = h.disable(); }
     if let Some(h) = HOOK_KEYBD_EVENT.get()    { let _ = h.disable(); }
